+++ docs/examples/03_application_pinns.py
@@ -194,7 +194,7 @@ X_boundary = sample_boundary()
 
 
 # Function that computes the neural network's Laplacian
-lap_f = laplacian(f, zeros(2, dtype=DTYPE))  # uses collapsed Taylor mode
+lap_f = laplacian(f, (zeros(2, dtype=DTYPE),))  # uses collapsed Taylor mode
 # Capture the operator's graph so we can apply CSE + dead code elimination.
 lap_f, _ = capture_graph(lap_f, (zeros(2, dtype=DTYPE),))
 common_subexpression_elimination(lap_f.graph)